separate-billing/requirements.txt
oslo.config==4.7.0
oslo.log==3.21.0
oslo.utils==3.26.0
python-keystoneclient==3.10.0
python-distilclient>=0.3.2
retrying>=1.2.3,!=1.3.0
prettytable>=0.7.2
